+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Note that since we don't clearly distinguish between a public and private interf
   - Support `surface_type` (`molecular` / `gaussian`) on for `surface` representation nodes
   - Add `instance` node type
   - Support transforming and instancing of structures, components, and volumes
+  - Use params hash for node version for more performant tree diffs
 - Added new color schemes, synchronized with D3.js ('inferno', 'magma', 'turbo', 'rainbow', 'sinebow', 'warm', 'cool', 'cubehelix-default', 'category-10', 'observable-10', 'tableau-10')
 - Snapshot Markdown improvements
   - Add `MarkdownExtensionManager` (`PluginContext.managers.markdownExtensions`)
@@ -65,6 +66,7 @@ Note that since we don't clearly distinguish between a public and private interf
     - Refactor `Canvas3dInteractionHelper` internals to use async picking for move events
 - Add `enable` param for post-processing effects. If false, no effects are applied.
 - Change the parsing of residue names in PDB files from 3-letter to 4-letter.
+- Support versioning transform using a hash function in `mol-state`
 
 ## [v4.18.0] - 2025-06-08
 - MolViewSpec extension: